move files, nested modal content and arrow function for content
